requirements.txt
django
google-generativeai
python-dotenv
requests
gunicorn
whitenoise
dj-database-url
